requirements.txt
openai>=0.27.0
PyMySQL>=1.0.2
pymupdf>=1.22.0
python-dotenv>=1.0.0
tqdm>=4.0.0